support only etcd

(cherry picked from commit ea259db913552fd3a124aaa642615a8d1c4a0717)
